# bff-server/swagger.yaml
openapi: 3.0.3
info:
  title: Moodle BFF API
  description: Backend For Frontend API for Moodle integration
  version: 1.0.0
  contact:
    name: Moodle BFF API Support
servers:
  - url: https://15.152.220.38
    description: Production server (HTTPS via nginx)
  - url: http://15.152.220.38
    description: Development server (HTTP, redirects to HTTPS)
  - url: http://localhost:3001
    description: Local development (direct access)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionId
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
...
    Course:
      type: object
      properties:
        id:
          type: integer
        fullname:
          type: string
        shortname:
          type: string
        categoryid:
          type: integer
        summary:
          type: string
        startdate:
          type: integer
        enddate:
          type: integer
        visible:
          type: boolean
        progress:
          type: integer
          description: コースの進捗率（0-100%）。アクティビティの完了状況から計算されます。
          example: 75
          minimum: 0
          maximum: 100
...
    ProfileUpdate:
      type: object
      properties:
        nick_name:
          type: string
          maxLength: 256
          nullable: true
          description: ニックネーム
          example: "たろう"
        self_intro:
          type: string
          nullable: true
          description: 自己紹介文
          example: "プログラミングを学んでいる学生です"
        target_job:
          type: string
          maxLength: 256
          nullable: true
          description: なりたい職業・肩書き
          example: "Webエンジニア"
        ideal_work_style:
          type: string
          maxLength: 256
          nullable: true
          description: 理想の働き方
          example: "リモートワーク中心"
        monthly_goal:
          type: string
          maxLength: 256
          nullable: true
          description: 今月の目標
          example: "JavaScriptの基礎を完了する"
        goal:
          type: string
          nullable: true
          description: 目標
          example: "フルスタックエンジニアになる"
        badge_count:
          type: integer
          minimum: 0
          nullable: true
          description: 獲得バッジ数
          example: 5
    Profile:
      type: object
      required:
        - mdl_user_id
      properties:
        mdl_user_id:
          type: integer
          description: Moodle User ID
          example: 2
        nick_name:
          type: string
          nullable: true
          description: ニックネーム
          example: "たろう"
        self_intro:
          type: string
          nullable: true
          description: 自己紹介文
          example: "プログラミングを学んでいる学生です"
        target_job:
          type: string
          nullable: true
          description: なりたい職業・肩書き
          example: "Webエンジニア"
        ideal_work_style:
          type: string
          nullable: true
          description: 理想の働き方
          example: "リモートワーク中心"
        monthly_goal:
          type: string
          nullable: true
          description: 今月の目標
          example: "JavaScriptの基礎を完了する"
        goal:
          type: string
          nullable: true
          description: 目標
          example: "フルスタックエンジニアになる"
        badge_count:
          type: integer
          nullable: true
          default: 0
          description: 獲得バッジ数
          example: 5
    ResumeCourse:
      type: object
      properties:
        courseid:
          type: integer
          description: コースID
          example: 2
        coursename:
          type: string
          description: コース名
          example: "Introduction to Programming"
        fullname:
          type: string
          description: コースのフルネーム
          example: "Introduction to Programming"
        shortname:
          type: string
          description: コースの短縮名
          example: "PROG101"
        summary:
          type: string
          description: コースの概要
          example: "Learn the basics of programming"
        progress:
          type: integer
          description: コースの進捗率（0-100%）。アクティビティの完了状況から計算されます。
          example: 65
          minimum: 0
          maximum: 100
        last_access_time:
          type: string
          format: date-time
          description: 最終アクセス日時
          example: "2024-01-28T10:30:00Z"
...
paths:
...
  /api/moodle/courses:
    get:
      tags:
        - Moodle Courses
      summary: Get all courses
      description: Get all courses in the Moodle system
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of all courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/moodle/courses/{userid}:
    get:
      tags:
        - Moodle Courses
      summary: Get enrolled courses by user ID
      description: Get all courses the specified user is enrolled in, including progress calculated from activity completion status
      security:
        - cookieAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: List of enrolled courses with progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Course'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/webcoach/profile/{userid}:
    get:
      tags:
        - WebCoach
      summary: Get user profile
      description: Get WebCoach profile information for a user from the API server
      security:
        - cookieAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - can only access own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to get profile from API server
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - WebCoach
      summary: Update user profile
      description: |
        Update WebCoach profile information. This endpoint forwards the request to the API server.

        All fields are optional. Only the fields you want to update need to be included in the request.
      security:
        - cookieAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match the authenticated user's ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdate'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - can only update own profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to update profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/webcoach/resumecourse/{userid}:
    get:
      tags:
        - WebCoach
      summary: Get resume courses
      description: Get recommended courses to resume for a user with progress calculated from activity completion status
      security:
        - cookieAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        '200':
          description: List of courses to resume with progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResumeCourse'
    post:
      tags:
        - WebCoach
      summary: Update resume course data
      description: Update or create resume course information for a user
      security:
        - cookieAuth: []
      parameters:
        - name: userid
          in: path
          required: true
          schema:
            type: integer
          description: User ID (must match the authenticated user's ID)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - courseid
                - progress_percent
              properties:
                courseid:
                  type: integer
                  description: Moodle course ID
                  example: 1
                progress_percent:
                  type: integer
                  minimum: 0
                  maximum: 100
                  description: Course completion percentage (0-100)
                  example: 0
      responses:
        '200':
          description: Resume course data updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResumeCourse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - can only update own data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to update resume course data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...